+++ extensions/google-calendar/src/lib/google.ts
@@ -1,5 +1,5 @@
 import { auth, calendar_v3 } from "@googleapis/calendar";
-import { OAuthService, useCachedPromise, withAccessToken, withCache } from "@raycast/utils";
+import { OAuthService, useCachedPromise, withAccessToken } from "@raycast/utils";
 import { people_v1 } from "@googleapis/people";
 import { Tool } from "@raycast/api";
 import { getClientId } from "./utils";
@@ -8,7 +8,9 @@ let calendar: calendar_v3.Calendar | null = null;
 let people: people_v1.People | null = null;
 
 export const GOOGLE_OAUTH_SCOPES = [
-  "https://www.googleapis.com/auth/calendar",
+  "https://www.googleapis.com/auth/calendar.events",
+  "https://www.googleapis.com/auth/calendar.calendarlist.readonly",
+  "https://www.googleapis.com/auth/calendar.freebusy",
   "https://www.googleapis.com/auth/contacts.readonly",
   "https://www.googleapis.com/auth/contacts.other.readonly",
   "https://www.googleapis.com/auth/userinfo.email",
@@ -85,29 +87,10 @@ export async function searchContacts(query?: string) {
   return [...(contacts ?? []), ...(otherContacts ?? [])];
 }
 
-export async function getAutoAddHangouts() {
-  const cachedFunction = withCache(async () => {
-    const calendar = getCalendarClient();
-    const settings = await calendar.settings.get({ setting: "autoAddHangouts" });
-    return settings.data.value === "true";
-  });
-  return await cachedFunction();
-}
-
 export function useContacts(query?: string) {
   return useCachedPromise(searchContacts, [query], {
     keepPreviousData: true,
   });
-}
-
-export function useCalendar(calendarId: string) {
-  return useCachedPromise(
-    async (calendarId: string) => {
-      const calendar = getCalendarClient();
-      return await calendar.calendars.get({ calendarId });
-    },
-    [calendarId],
-  );
 }
 
 export function useCalendars() {